Refactor password UI into separate PasswordDialog class
- Extract password dialog UI to dedicated PasswordDialog class
- Reduce ClientUIManager from 1488 to 1155 lines (333 lines saved)
- Remove duplicate GetUIManager() and GetPlayerManager() methods
- Use static properties directly (UIManager, Players) instead of getter methods
- Improves separation of concerns and maintainability
